+++ src/any_llm/__init__.py
@@ -8,8 +8,8 @@ from any_llm.api import (
     list_models_async,
     responses,
 )
+from any_llm.constants import ProviderName
 from any_llm.exceptions import MissingApiKeyError
-from any_llm.provider import ProviderName
 from any_llm.tools import callable_to_tool, prepare_tools
 
 __all__ = [
