+++ src/any_llm/api.py
@@ -3,12 +3,25 @@ from typing import Any, Literal
 
 from pydantic import BaseModel
 
-from any_llm.provider import ClientConfig, ProviderFactory, ProviderName
+from any_llm.config import ClientConfig
+from any_llm.constants import ProviderName
+from any_llm.factory import ProviderFactory
 from any_llm.tools import prepare_tools
 from any_llm.types.completion import ChatCompletion, ChatCompletionChunk, ChatCompletionMessage, CreateEmbeddingResponse
 from any_llm.types.model import Model
 from any_llm.types.responses import Response, ResponseInputParam, ResponseStreamEvent
 from any_llm.utils.api import _process_completion_params
+
+# from any_llm import AnyLLM, LLMProvider, LLMConfig
+
+# llm = AnyLLM.create(
+#     LLMProvider.MISTRAL,
+#     LLMConfig(
+#         api_key=api_key,
+#         api_base=api_base,
+#         http_client=http_client
+#     )
+# )
 
 
 def completion(
