+++ src/any_llm/provider.py
@@ -1,19 +1,13 @@
 # Inspired by https://github.com/andrewyng/aisuite/tree/main/aisuite
 import asyncio
-import builtins
-import importlib
-import logging
 import os
-import warnings
 from abc import ABC, abstractmethod
 from collections.abc import AsyncIterator, Iterator, Sequence
-from enum import StrEnum
-from pathlib import Path
 from typing import Any
 
-from pydantic import BaseModel
-
-from any_llm.exceptions import MissingApiKeyError, UnsupportedProviderError
+from any_llm.config import ClientConfig
+from any_llm.constants import INSIDE_NOTEBOOK
+from any_llm.exceptions import MissingApiKeyError
 from any_llm.types.completion import (
     ChatCompletion,
     ChatCompletionChunk,
@@ -24,68 +18,6 @@ from any_llm.types.model import Model
 from any_llm.types.provider import ProviderMetadata
 from any_llm.types.responses import Response, ResponseInputParam, ResponseStreamEvent
 from any_llm.utils.aio import async_iter_to_sync_iter, run_async_in_sync
-
-logger = logging.getLogger(__name__)
-
-
-INSIDE_NOTEBOOK = hasattr(builtins, "__IPYTHON__")
-
-
-class ProviderName(StrEnum):
-    """String enum for supported providers."""
-
-    ANTHROPIC = "anthropic"
-    BEDROCK = "bedrock"
-    AZURE = "azure"
-    AZUREOPENAI = "azureopenai"
-    CEREBRAS = "cerebras"
-    COHERE = "cohere"
-    DATABRICKS = "databricks"
-    DEEPSEEK = "deepseek"
-    FIREWORKS = "fireworks"
-    GEMINI = "gemini"
-    GROQ = "groq"
-    HUGGINGFACE = "huggingface"
-    INCEPTION = "inception"
-    LLAMA = "llama"
-    LMSTUDIO = "lmstudio"
-    LLAMAFILE = "llamafile"
-    LLAMACPP = "llamacpp"
-    MISTRAL = "mistral"
-    MOONSHOT = "moonshot"
-    NEBIUS = "nebius"
-    OLLAMA = "ollama"
-    OPENAI = "openai"
-    OPENROUTER = "openrouter"
-    PORTKEY = "portkey"
-    SAMBANOVA = "sambanova"
-    SAGEMAKER = "sagemaker"
-    TOGETHER = "together"
-    VERTEXAI = "vertexai"
-    VOYAGE = "voyage"
-    WATSONX = "watsonx"
-    XAI = "xai"
-    PERPLEXITY = "perplexity"
-
-    @classmethod
-    def from_string(cls, value: "str | ProviderName") -> "ProviderName":
-        if isinstance(value, cls):
-            return value
-
-        formatted_value = value.strip().lower()
-        try:
-            return cls(formatted_value)
-        except ValueError as exc:
-            supported = [provider.value for provider in cls]
-            raise UnsupportedProviderError(value, supported) from exc
-
-
-class ClientConfig(BaseModel):
-    """Configuration for the underlying client used by the provider."""
-
-    api_key: str | None = None
-    api_base: str | None = None
-    client_args: dict[str, Any] | None = None
 
 
 class Provider(ABC):
@@ -292,119 +224,3 @@ class Provider(ABC):
 
     async def list_models_async(self, **kwargs: Any) -> Sequence[Model]:
         return await asyncio.to_thread(self.list_models, **kwargs)
-
-
-class ProviderFactory:
-    """Factory to dynamically load provider instances based on the naming conventions."""
-
-    PROVIDERS_DIR = Path(__file__).parent / "providers"
-
-    @classmethod
-    def create_provider(cls, provider_key: str | ProviderName, config: ClientConfig) -> Provider:
-        """Dynamically load and create an instance of a provider based on the naming convention."""
-        provider_key = ProviderName.from_string(provider_key).value
-
-        provider_class_name = f"{provider_key.capitalize()}Provider"
-        provider_module_name = f"{provider_key}"
-
-        module_path = f"any_llm.providers.{provider_module_name}"
-
-        try:
-            module = importlib.import_module(module_path)
-        except ImportError as e:
-            msg = f"Could not import module {module_path}: {e!s}. Please ensure the provider is supported by doing ProviderFactory.get_supported_providers()"
-            raise ImportError(msg) from e
-
-        provider_class: type[Provider] = getattr(module, provider_class_name)
-        return provider_class(config=config)
-
-    @classmethod
-    def get_provider_class(cls, provider_key: str | ProviderName) -> type[Provider]:
-        """Get the provider class without instantiating it.
-
-        Args:
-            provider_key: The provider key (e.g., 'anthropic', 'openai')
-
-        Returns:
-            The provider class
-        """
-        provider_key = ProviderName.from_string(provider_key).value
-
-        provider_class_name = f"{provider_key.capitalize()}Provider"
-        provider_module_name = f"{provider_key}"
-
-        module_path = f"any_llm.providers.{provider_module_name}"
-
-        try:
-            module = importlib.import_module(module_path)
-        except ImportError as e:
-            msg = f"Could not import module {module_path}: {e!s}. Please ensure the provider is supported by doing ProviderFactory.get_supported_providers()"
-            raise ImportError(msg) from e
-
-        provider_class: type[Provider] = getattr(module, provider_class_name)
-        return provider_class
-
-    @classmethod
-    def get_supported_providers(cls) -> list[str]:
-        """Get a list of supported provider keys."""
-        return [provider.value for provider in ProviderName]
-
-    @classmethod
-    def get_all_provider_metadata(cls) -> list[ProviderMetadata]:
-        """Get metadata for all supported providers.
-
-        Returns:
-            List of dictionaries containing provider metadata
-        """
-        providers: list[ProviderMetadata] = []
-        for provider_key in cls.get_supported_providers():
-            provider_class = cls.get_provider_class(provider_key)
-            metadata = provider_class.get_provider_metadata()
-            providers.append(metadata)
-
-        # Sort providers by name
-        providers.sort(key=lambda x: x.name)
-        return providers
-
-    @classmethod
-    def get_provider_enum(cls, provider_key: str) -> ProviderName:
-        """Convert a string provider key to a ProviderName enum."""
-        try:
-            return ProviderName(provider_key)
-        except ValueError as e:
-            supported = [provider.value for provider in ProviderName]
-            raise UnsupportedProviderError(provider_key, supported) from e
-
-    @classmethod
-    def split_model_provider(cls, model: str) -> tuple[ProviderName, str]:
-        """Extract the provider key from the model identifier.
-
-        Supports both new format 'provider:model' (e.g., 'mistral:mistral-small')
-        and legacy format 'provider/model' (e.g., 'mistral/mistral-small').
-
-        The legacy format will be deprecated in version 1.0.
-        """
-        colon_index = model.find(":")
-        slash_index = model.find("/")
-
-        # Determine which delimiter comes first
-        if colon_index != -1 and (slash_index == -1 or colon_index < slash_index):
-            # The colon came first, so it's using the new syntax.
-            provider, model_name = model.split(":", 1)
-        elif slash_index != -1:
-            # Slash comes first, so it's the legacy syntax
-            warnings.warn(
-                f"Model format 'provider/model' is deprecated and will be removed in version 1.0. "
-                f"Please use 'provider:model' format instead. Got: '{model}'",
-                DeprecationWarning,
-                stacklevel=3,
-            )
-            provider, model_name = model.split("/", 1)
-        else:
-            msg = f"Invalid model format. Expected 'provider:model' or 'provider/model', got '{model}'"
-            raise ValueError(msg)
-
-        if not provider or not model_name:
-            msg = f"Invalid model format. Expected 'provider:model' or 'provider/model', got '{model}'"
-            raise ValueError(msg)
-        return cls.get_provider_enum(provider), model_name
