+++ src/any_llm/providers/vertexai/vertexai.py
@@ -1,8 +1,8 @@
 import os
 from typing import TYPE_CHECKING
 
+from any_llm.config import ClientConfig
 from any_llm.exceptions import MissingApiKeyError
-from any_llm.provider import ClientConfig
 from any_llm.providers.gemini.base import GoogleProvider
 
 if TYPE_CHECKING:
