+++ tests/constants.py
@@ -1,6 +1,6 @@
 import os
 
-from any_llm.provider import ProviderName
+from any_llm.constants import ProviderName
 
 LOCAL_PROVIDERS = [ProviderName.LLAMACPP, ProviderName.OLLAMA, ProviderName.LMSTUDIO, ProviderName.LLAMAFILE]
