+++ tests/integration/test_completion.py
@@ -10,7 +10,7 @@ from openai import APIConnectionError
 
 from any_llm import ProviderName, acompletion
 from any_llm.exceptions import MissingApiKeyError
-from any_llm.provider import ProviderFactory
+from any_llm.factory import ProviderFactory
 from any_llm.types.completion import ChatCompletion, ChatCompletionMessage
 from tests.constants import EXPECTED_PROVIDERS, LOCAL_PROVIDERS
