+++ tests/integration/test_tool.py
@@ -7,7 +7,7 @@ from openai import APIConnectionError
 
 from any_llm import ProviderName, acompletion
 from any_llm.exceptions import MissingApiKeyError
-from any_llm.provider import ProviderFactory
+from any_llm.factory import ProviderFactory
 from tests.constants import EXPECTED_PROVIDERS, LOCAL_PROVIDERS
 
 if TYPE_CHECKING:
