+++ tests/unit/providers/test_xai_provider.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from any_llm.provider import ClientConfig
+from any_llm.config import ClientConfig
 from any_llm.types.completion import ChatCompletion, CompletionParams
